+++ packages/xchain-mayachain/src/const.ts
@@ -7,7 +7,7 @@ const DEFAULT_EXPLORER_URL = 'https://explorer.mayachain.info'
 const txUrl = `${DEFAULT_EXPLORER_URL}/tx`
 const addressUrl = `${DEFAULT_EXPLORER_URL}/address`
 
-export const DECIMAL = 8
+export const CACAO_DECIMAL = 8
 export const DEFAULT_GAS_ADJUSTMENT = 2
 export const DEFAULT_GAS_LIMIT_VALUE = '4000000'
 export const DEPOSIT_GAS_LIMIT_VALUE = '600000000'