+++ packages/xchain-thorchain/src/utils.ts
@@ -16,7 +16,7 @@ import axios from 'axios'
 import * as bech32Buffer from 'bech32-buffer'
 import Long from 'long'
 
-import { AssetRuneNative, DECIMAL, DEFAULT_GAS_ADJUSTMENT } from './const'
+import { AssetRuneNative, DEFAULT_GAS_ADJUSTMENT, RUNE_DECIMAL } from './const'
 import { ChainId, ExplorerUrls, NodeInfoResponse, TxData } from './types'
 import { MsgNativeTx } from './types/messages'
 import types from './types/proto/MsgCompiled'
@@ -121,11 +121,11 @@ export const getDepositTxDataFromLogs = (
   const transferDataList: TransferDataList = events.reduce((acc: TransferDataList, { type, attributes }) => {
     if (type === 'transfer') {
       return attributes.reduce((acc2, { key, value }, index) => {
-        if (index % 3 === 0) acc2.push({ sender: '', recipient: '', amount: baseAmount(0, DECIMAL) })
+        if (index % 3 === 0) acc2.push({ sender: '', recipient: '', amount: baseAmount(0, RUNE_DECIMAL) })
         const newData = acc2[acc2.length - 1]
         if (key === 'sender') newData.sender = value
         if (key === 'recipient') newData.recipient = value
-        if (key === 'amount') newData.amount = baseAmount(value.replace(/rune/, ''), DECIMAL)
+        if (key === 'amount') newData.amount = baseAmount(value.replace(/rune/, ''), RUNE_DECIMAL)
         return acc2
       }, acc)
     }
@@ -153,7 +153,7 @@ export const getDepositTxDataFromLogs = (
  * @returns {Fees} The default fee.
  */
 export const getDefaultFees = (): Fees => {
-  const fee = assetToBase(assetAmount(0.02 /* 0.02 RUNE */, DECIMAL))
+  const fee = assetToBase(assetAmount(0.02 /* 0.02 RUNE */, RUNE_DECIMAL))
   return singleFee(FeeType.FlatFee, fee)
 }
 
@@ -383,7 +383,7 @@ export const getBalance = async ({
   return balances
     .map((balance) => ({
       asset: (balance.denom && assetFromDenom(balance.denom)) || AssetRuneNative,
-      amount: baseAmount(balance.amount, DECIMAL),
+      amount: baseAmount(balance.amount, RUNE_DECIMAL),
     }))
     .filter(
       (balance) => !assets || assets.filter((asset) => assetToString(balance.asset) === assetToString(asset)).length,