+++ packages/xchain-bitcoincash/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.16.1 (2023-05-18)
+
+## Add
+
+- New client function getAssetInfo() returns chain, decimals and asset
+- renamed defaultBCHParams to defaultBchParams to be more consistent with CamelCase
+
 # v0.16.0 (2023-05-02)
 
 ## Update