+++ packages/xchain-cosmos/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.21.1 (2023-05-18)
+
+## Add
+
+- New client function getAssetInfo() returns chain, decimals and asset
+
 # v0.21.0 (2023-05-02)
 
 ## Update