+++ packages/xchain-evm/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.2.2 (2023-05-18)
+
+## Add
+
+- New client function getAssetInfo() returns chain, decimals and asset
+
 # v0.2.1 (2023-05-09)
 
 ## Update